Update Run.gd to consider vertical camera angle
This is a relatively small change, but it fixes an interesting "issue" if you allow the camera to be rotated vertically. Without this, and if the camera is rotated vertically, the character will start to jitter and slow down while they run or sprint. 

This is because the angle of the camera has now changed vertically, so the vector is in a different place; the signed_angle_to function does not properly account for this, and thus, the humanoid velocity gets wonky as a result (animator speed scale is also affected by this, but it is not the main culprit).

By sliding the input_direction onto y-axis (Vector3.UP), we project the camera angle onto the same axis associated with character's face direction. So both vectors end up on the same axis, and thus, signed_angle_to gives the correct result again.